feat: url 선택 가능한 사이드바 생성
