+++ gitlio/app/(edit)/new/showproject/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function page() {
+  return <div>page</div>;
+}
+
+export default page;
